Isolate adventure code to adventure_anywhere package (#6)
* Instantiate game engine in use case rather than passing thru context

This makes more sense to me because:
1) The 'Game' object isn't really a "plugin" in the same way as the
saves gateway that does i/o
2) A gateway plugin should be able to be created once by the delivery
layer and repeatedly passed to the use case if that's how delivery is
setup (e.g. a long-running server created a db connection and reuses
that). The 'game gateway' was stateful and had to be recreated each
time the use case is invoked.
3) my only weird feeling is having a stateful object created in the
use case

* Rename 'out' to 'violation'

* Move delivery-specific config into context.config object

* Create adventure_anywhere package and move sms to sms_adventure

* Update doctests